fix mobile styling for search events page
