+++ src/pages/SearchEventsCard.module.css
@@ -18,6 +18,7 @@
   transition: var(--card-transition);
   background: var(--secondary-clr);
   text-decoration: none;
+  overflow: hidden;
 }
 
 .card:hover {
@@ -27,10 +28,6 @@
   -moz-transform: translateY(var(--card-shift-y));
   -ms-transform: translateY(var(--card-shift-y));
   -o-transform: translateY(var(--card-shift-y));
-}
-
-.card:hover .arrow svg {
-  transform: translate(4px, -4px);
 }
 
 .card img {
@@ -46,8 +43,8 @@
   flex-direction: column;
   justify-content: space-between;
   padding: 10px;
+  height: 14em;
   overflow: hidden;
-  height: 13em;
   background: var(--shadow-gradient);
   border-bottom-left-radius: var(--card-border-radius);
   border-bottom-right-radius: var(--card-border-radius);
@@ -64,15 +61,6 @@
   margin-bottom: var(--card-text-margin);
   text-wrap: wrap;
   overflow: hidden;
-}
-
-.arrow svg {
-  position: relative;
-  width: 18px;
-  height: 18px;
-  stroke: white;
-  left: 3px;
-  transition: var(--card-transition);
 }
 
 .location {
@@ -94,26 +82,19 @@
     flex: 0 0 calc(50% - var(--card-gap));
     max-width: calc(50% - var(--card-gap));
   }
+
+  .title {
+    font-size: 20px;
+  }
 }
 
 @media screen and (max-width: 980px) {
   .card {
-    flex-direction: row;
     flex: 0 0 100%;
     max-width: 100%;
-    height: 13rem;
-    overflow: hidden;
   }
 
-  .card img {
-    width: 13rem;
-    height: 100%;
-    object-position: top left;
-    border-top-right-radius: 0;
-    border-bottom-left-radius: var(--card-border-radius);
-  }
-
-  .title {
-    font-size: 16px;
+  .content {
+    height: auto;
   }
 }
